worker/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
onnxruntime-gpu>=1.16.0
